+++ .github/workflows/examples.yml
@@ -12,31 +12,24 @@ jobs:
     name: Build SQLx CLI
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v4
 
-      - uses: actions-rs/toolchain@v1
-        with:
-          profile: minimal
-          toolchain: stable
-          override: true
+      - name: Use latest Rust
+        run: rustup override set stable
 
-      - uses: Swatinem/rust-cache@v1
+      - uses: Swatinem/rust-cache@v2
         with:
           key: sqlx-cli
 
-      - uses: actions-rs/cargo@v1
-        with:
-          command: build
-          args: >
+      - run: >
+            cargo build
             -p sqlx-cli
             --bin sqlx
             --release
             --no-default-features
             --features mysql,postgres,sqlite
-        env:
-          RUSTFLAGS: -D warnings
 
-      - uses: actions/upload-artifact@v3
+      - uses: actions/upload-artifact@v4
         with:
           name: sqlx-cli
           path: target/release/sqlx
@@ -56,7 +49,7 @@ jobs:
 
     steps:
       - name: Get SQLx-CLI
-        uses: actions/download-artifact@v3
+        uses: actions/download-artifact@v4
         with:
           name: sqlx-cli
           # $HOME is interpreted differently by the shell
@@ -68,15 +61,9 @@ jobs:
           echo /home/runner/.local/bin >> $GITHUB_PATH
           sleep 10
 
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v4
 
-      - uses: actions-rs/toolchain@v1
-        with:
-          profile: minimal
-          toolchain: stable
-          override: true
-
-      - uses: Swatinem/rust-cache@v1
+      - uses: Swatinem/rust-cache@v2
         with:
           key: mysql-examples
 
@@ -87,13 +74,9 @@ jobs:
         run: sqlx db setup
 
       - name: Todos (Run)
-        uses: actions-rs/cargo@v1
         env:
           DATABASE_URL: mysql://root:password@localhost:3306/todos?ssl-mode=disabled
-        with:
-          # TODO: test full CLI
-          command: run
-          args: -p sqlx-example-mysql-todos
+        run: cargo run -p sqlx-example-mysql-todos
 
   postgres:
     name: PostgreSQL Examples
@@ -110,7 +93,7 @@ jobs:
 
     steps:
       - name: Get SQLx-CLI
-        uses: actions/download-artifact@v3
+        uses: actions/download-artifact@v4
         with:
           name: sqlx-cli
           path: /home/runner/.local/bin
@@ -121,15 +104,9 @@ jobs:
           echo $HOME/.local/bin >> $GITHUB_PATH
           sleep 10
 
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v4
 
-      - uses: actions-rs/toolchain@v1
-        with:
-          profile: minimal
-          toolchain: stable
-          override: true
-
-      - uses: Swatinem/rust-cache@v1
+      - uses: Swatinem/rust-cache@v2
         with:
           key: pg-examples
 
@@ -140,28 +117,19 @@ jobs:
         run: sqlx db setup
 
       - name: Axum Social with Tests (Check)
-        uses: actions-rs/cargo@v1
         env:
           DATABASE_URL: postgres://postgres:password@localhost:5432/axum-social
-        with:
-          command: check
-          args: -p sqlx-example-postgres-axum-social
+        run: cargo check -p sqlx-example-postgres-axum-social
 
       - name: Axum Social with Tests (Test)
-        uses: actions-rs/cargo@v1
         env:
           DATABASE_URL: postgres://postgres:password@localhost:5432/axum-social
-        with:
-          command: test
-          args: -p sqlx-example-postgres-axum-social
+        run: cargo test -p sqlx-example-postgres-axum-social
 
       # The Chat example has an interactive TUI which is not trivial to test automatically,
       # so we only check that it compiles.
       - name: Chat (Check)
-        uses: actions-rs/cargo@v1
-        with:
-          command: check
-          args: -p sqlx-example-postgres-chat
+        run: cargo check -p sqlx-example-postgres-chat
 
       - name: Files (Setup)
         working-directory: examples/postgres/files
@@ -170,12 +138,9 @@ jobs:
         run: sqlx db setup
 
       - name: Files (Run)
-        uses: actions-rs/cargo@v1
         env:
           DATABASE_URL: postgres://postgres:password@localhost:5432/files
-        with:
-          command: run
-          args: -p sqlx-example-postgres-files
+        run: cargo run -p sqlx-example-postgres-files
 
       - name: JSON (Setup)
         working-directory: examples/postgres/json
@@ -184,12 +149,9 @@ jobs:
         run: sqlx db setup
 
       - name: JSON (Run)
-        uses: actions-rs/cargo@v1
         env:
           DATABASE_URL: postgres://postgres:password@localhost:5432/json
-        with:
-          command: run
-          args: -p sqlx-example-postgres-json
+        run: cargo run -p sqlx-example-postgres-json
 
       - name: Listen (Setup)
         working-directory: examples/postgres/listen
@@ -198,12 +160,9 @@ jobs:
         run: sqlx db create
 
       - name: Listen (Run)
-        uses: actions-rs/cargo@v1
         env:
           DATABASE_URL: postgres://postgres:password@localhost:5432/listen
-        with:
-          command: run
-          args: -p sqlx-example-postgres-listen
+        run: cargo run -p sqlx-example-postgres-listen
 
       - name: Mockable TODOs (Setup)
         working-directory: examples/postgres/mockable-todos
@@ -212,13 +171,9 @@ jobs:
         run: sqlx db setup
 
       - name: Mockable TODOs (Run)
-        uses: actions-rs/cargo@v1
         env:
           DATABASE_URL: postgres://postgres:password@localhost:5432/mockable-todos
-        with:
-          # TODO: test full CLI
-          command: run
-          args: -p sqlx-example-postgres-mockable-todos
+        run: cargo run -p sqlx-example-postgres-mockable-todos
 
       - name: TODOs (Setup)
         working-directory: examples/postgres/todos
@@ -227,13 +182,10 @@ jobs:
         run: sqlx db setup
 
       - name: TODOs (Run)
-        uses: actions-rs/cargo@v1
         env:
           DATABASE_URL: postgres://postgres:password@localhost:5432/todos
-        with:
-          # TODO: test full CLI
-          command: run
-          args: -p sqlx-example-postgres-todos
+        # TODO: test full CLI
+        run: cargo run -p sqlx-example-postgres-todos
 
       - name: Transaction (Setup)
         working-directory: examples/postgres/transaction
@@ -242,12 +194,9 @@ jobs:
         run: sqlx db setup
 
       - name: Transaction (Run)
-        uses: actions-rs/cargo@v1
         env:
           DATABASE_URL: postgres://postgres:password@localhost:5432/txn
-        with:
-          command: run
-          args: -p sqlx-example-postgres-transaction
+        run: cargo run -p sqlx-example-postgres-transaction
 
   sqlite:
     name: SQLite Examples
@@ -256,7 +205,7 @@ jobs:
 
     steps:
       - name: Get SQLx-CLI
-        uses: actions/download-artifact@v3
+        uses: actions/download-artifact@v4
         with:
           name: sqlx-cli
           path: /home/runner/.local/bin
@@ -266,15 +215,9 @@ jobs:
           chmod +x /home/runner/.local/bin/sqlx
           echo /home/runner/.local/bin >> $GITHUB_PATH
 
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v4
 
-      - uses: actions-rs/toolchain@v1
-        with:
-          profile: minimal
-          toolchain: stable
-          override: true
-
-      - uses: Swatinem/rust-cache@v1
+      - uses: Swatinem/rust-cache@v2
         with:
           key: sqlite-examples
 
@@ -284,9 +227,6 @@ jobs:
         run: sqlx db setup --source=examples/sqlite/todos/migrations
 
       - name: TODOs (Run)
-        uses: actions-rs/cargo@v1
         env:
           DATABASE_URL: sqlite://todos.sqlite
-        with:
-          command: run
-          args: -p sqlx-example-sqlite-todos
+        run: cargo run -p sqlx-example-sqlite-todos
