+++ .github/workflows/sqlx.yml
@@ -79,52 +79,6 @@ jobs:
           --manifest-path sqlx-core/Cargo.toml
           --features json,_rt-${{ matrix.runtime }},_tls-${{ matrix.tls }}
 
-  cli-test:
-    name: CLI Unit Test
-    runs-on: ubuntu-22.04
-    steps:
-      - uses: actions/checkout@v4
-
-      - uses: Swatinem/rust-cache@v2
-        with:
-          key: ${{ runner.os }}-test
-
-      - run: cargo test --manifest-path sqlx-cli/Cargo.toml
-
-  cli:
-    name: CLI Binaries
-    runs-on: ${{ matrix.os }}
-
-    strategy:
-      matrix:
-        os: [ubuntu-latest, windows-latest] #, macOS-latest]
-        include:
-          - os: ubuntu-latest
-            target: x86_64-unknown-linux-musl
-            args: --features openssl-vendored
-            bin: target/debug/cargo-sqlx
-          - os: windows-latest
-            target: x86_64-pc-windows-msvc
-            bin: target/debug/cargo-sqlx.exe
-          # FIXME: macOS build fails because of missing pin-project-internal
-    #          - os: macOS-latest
-    #            target: x86_64-apple-darwin
-    #            bin: target/debug/cargo-sqlx
-
-    steps:
-      - uses: actions/checkout@v4
-
-      - uses: Swatinem/rust-cache@v2
-        with:
-          key: ${{ runner.os }}-cli
-
-      - run: cargo build --manifest-path sqlx-cli/Cargo.toml --bin cargo-sqlx ${{ matrix.args }}
-
-      - uses: actions/upload-artifact@v2
-        with:
-          name: cargo-sqlx-${{ matrix.target }}
-          path: ${{ matrix.bin }}
-
   sqlite:
     name: SQLite
     runs-on: ubuntu-22.04
