+++ rust-toolchain.toml
@@ -2,5 +2,5 @@
 # We reserve the right to increase this version at any time without considering it to be a breaking change.
 # See the answer in FAQ.md for details.
 [toolchain]
-channel = "1.78"
+channel = "1.80"
 profile = "minimal"
