+++ sqlx-core/src/net/socket/buffered.rs
@@ -4,7 +4,7 @@ use std::{cmp, io};
 
 use crate::error::Error;
 
-use crate::io::{Decode, Encode};
+use crate::io::{AsyncRead, AsyncReadExt, Decode, Encode};
 
 // Tokio, async-std, and std all use this as the default capacity for their buffered I/O.
 const DEFAULT_BUF_SIZE: usize = 8192;
@@ -166,6 +166,25 @@ impl WriteBuffer {
         self.bytes_written = new_bytes_written;
 
         self.sanity_check();
+    }
+
+    /// Read into the buffer from `source`, returning the number of bytes read.
+    ///
+    /// The buffer is automatically advanced by the number of bytes read.
+    pub async fn read_from(&mut self, mut source: impl AsyncRead + Unpin) -> io::Result<usize> {
+        let read = match () {
+            // Tokio lets us read into the buffer without zeroing first
+            #[cfg(feature = "_rt-tokio")]
+            _ => source.read_buf(self.buf_mut()).await?,
+            #[cfg(not(feature = "_rt-tokio"))]
+            _ => source.read(self.init_remaining_mut()).await?,
+        };
+
+        if read > 0 {
+            self.advance(read);
+        }
+
+        Ok(read)
     }
 
     pub fn is_empty(&self) -> bool {
