+++ sqlx-macros-core/Cargo.toml
@@ -62,3 +62,8 @@ syn = { version = "2.0.52", default-features = false, features = ["full", "deriv
 tempfile = { version = "3.10.1" }
 quote = { version = "1.0.26", default-features = false }
 url = { version = "2.2.2", default-features = false }
+
+[lints.rust.unexpected_cfgs]
+level = "warn"
+# 1.80 will warn without this
+check-cfg = ['cfg(sqlx_macros_unstable)', 'cfg(procmacro2_semver_exempt)']
