+++ sqlx-macros-core/src/query/data.rs
@@ -98,7 +98,7 @@ impl DynQueryData {
             return Ok(cached);
         }
 
-        #[cfg(procmacr2_semver_exempt)]
+        #[cfg(procmacro2_semver_exempt)]
         {
             let path = path.as_ref().canonicalize()?;
             let path = path.to_str().ok_or_else(|| {
