+++ src/macros/mod.rs
@@ -52,19 +52,19 @@
 /// † Only callable if the query returns no columns; otherwise it's assumed the query *may* return at least one row.
 /// ## Requirements
 /// * The `DATABASE_URL` environment variable must be set at build-time to point to a database
-/// server with the schema that the query string will be checked against. All variants of `query!()`
-/// use [dotenv]<sup>1</sup> so this can be in a `.env` file instead.
+///   server with the schema that the query string will be checked against.
+///   All variants of `query!()` use [dotenv]<sup>1</sup> so this can be in a `.env` file instead.
 ///
 ///     * Or, `.sqlx` must exist at the workspace root. See [Offline Mode](#offline-mode-requires-the-offline-feature)
 ///       below.
 ///
 /// * The query must be a string literal, or concatenation of string literals using `+` (useful
-/// for queries generated by macro), or else it cannot be introspected (and thus cannot be dynamic
-/// or the result of another macro).
+///   for queries generated by macro), or else it cannot be introspected (and thus cannot be dynamic
+///   or the result of another macro).
 ///
 /// * The `QueryAs` instance will be bound to the same database type as `query!()` was compiled
-/// against (e.g. you cannot build against a Postgres database and then run the query against
-/// a MySQL database).
+///   against (e.g. you cannot build against a Postgres database and then run the query against
+///   a MySQL database).
 ///
 ///     * The schema of the database URL (e.g. `postgres://` or `mysql://`) will be used to
 ///       determine the database type.
@@ -426,7 +426,7 @@ macro_rules! query_file_unchecked (
 /// * The query must output at least one column.
 /// * The column names of the query must match the field names of the struct.
 /// * The field types must be the Rust equivalent of their SQL counterparts; see the corresponding
-/// module for your database for mappings:
+///   module for your database for mappings:
 ///     * Postgres: [crate::postgres::types]
 ///     * MySQL: [crate::mysql::types]
 ///         * Note: due to wire protocol limitations, the query macros do not know when
